+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Completions for `:RustLsp` subcommands' arguments.
+- DAP: Auto-generate source map, to allow stepping into `std`.
 
 ### Changed
 - Removed `plenary.nvim` dependency (`dap` and `quickfix` executor).